Adds to function parsing
Functions are now added to the symbol table with the correct index.
Prototype and Function creation should be alright in FunctionStatement.

Still missing is handling of the function parameters, handing off the
currently compiling function to the newly created one and the release back
at the end of the function body.

Also need to handle return statements correctly and there is no mechanism
for handling function calls right now, but this should be quite simple
once function parsing is in place.